SS-9007: Codebase improvements
