+++ components/shapediver/viewport/ViewportIcons.tsx
@@ -1,20 +1,21 @@
 import {OverlayPosition} from "@AppBuilderShared/components/shapediver/ui/OverlayWrapper";
-import {
-	ComponentContext,
-	renderViewportIcons,
-} from "@AppBuilderShared/context/ComponentContext";
+import {ComponentContext} from "@AppBuilderShared/context/ComponentContext";
 import {useViewportControls} from "@AppBuilderShared/hooks/shapediver/viewer/useViewportControls";
 import {useViewportId} from "@AppBuilderShared/hooks/shapediver/viewer/useViewportId";
 import {useShapeDiverStoreParameters} from "@AppBuilderShared/store/useShapeDiverStoreParameters";
 import {useShapeDiverStoreViewport} from "@AppBuilderShared/store/useShapeDiverStoreViewport";
 import {useShapeDiverViewportIconsStore} from "@AppBuilderShared/store/useShapeDiverViewportIconsStore";
-import {ButtonRenderContext} from "@AppBuilderShared/types/components/shapediver/componentTypes";
+import {
+	ButtonRenderContext,
+	getViewportIconComponent,
+} from "@AppBuilderShared/types/components/shapediver/componentTypes";
 import {ViewportTransparentBackgroundStyle} from "@AppBuilderShared/types/shapediver/viewport";
 import {
 	ViewportIconsOptionalProps,
 	ViewportIconsProps,
 } from "@AppBuilderShared/types/shapediver/viewportIcons";
-import {Paper, Transition, useProps} from "@mantine/core";
+import {ViewportIconButtonEnum} from "@AppBuilderShared/types/store/shapediverStoreViewportIcons";
+import {Divider, Paper, Transition, useProps} from "@mantine/core";
 import React, {useCallback, useContext, useMemo, useState} from "react";
 import {useShallow} from "zustand/react/shallow";
 import ViewportOverlayWrapper from "./ViewportOverlayWrapper";
@@ -196,12 +197,156 @@ export default function ViewportIcons(
 	);
 
 	const content = useMemo(() => {
-		return renderViewportIcons(
-			layout,
-			componentContext,
-			buttonContext,
-			dividerProps,
-		);
+		const renderButtonByKind = (
+			kind: ViewportIconButtonEnum,
+			componentContext: any,
+			buttonContext: any,
+		) => {
+			const {
+				viewport,
+				namespace,
+				buttonsDisabled,
+				executing,
+				hasPendingChanges,
+				iconsVisible,
+				fullscreenId,
+				...commonProps
+			} = buttonContext;
+
+			const ButtonComponent = getViewportIconComponent(
+				componentContext,
+				kind,
+			);
+			if (!ButtonComponent) return null;
+
+			switch (kind) {
+				case ViewportIconButtonEnum.Ar:
+					return React.createElement(ButtonComponent, {
+						key: "ar",
+						viewport,
+						...commonProps,
+					});
+				case ViewportIconButtonEnum.Zoom:
+					return React.createElement(ButtonComponent, {
+						key: "zoom",
+						viewport,
+						...commonProps,
+					});
+				case ViewportIconButtonEnum.Fullscreen:
+					return React.createElement(ButtonComponent, {
+						key: "fullscreen",
+						fullscreenId,
+						enableFullscreenBtn: true,
+						...commonProps,
+					});
+				case ViewportIconButtonEnum.Cameras:
+					return React.createElement(ButtonComponent, {
+						key: "cameras",
+						viewport,
+						visible: iconsVisible,
+						...commonProps,
+					});
+				case ViewportIconButtonEnum.Undo:
+					return React.createElement(ButtonComponent, {
+						key: "undo",
+						disabled:
+							buttonsDisabled || executing || hasPendingChanges,
+						hasPendingChanges,
+						executing,
+						...commonProps,
+					});
+				case ViewportIconButtonEnum.Redo:
+					return React.createElement(ButtonComponent, {
+						key: "redo",
+						disabled:
+							buttonsDisabled || executing || hasPendingChanges,
+						hasPendingChanges,
+						executing,
+						...commonProps,
+					});
+				case ViewportIconButtonEnum.Reload:
+					return React.createElement(ButtonComponent, {
+						key: "reload",
+						disabled:
+							!namespace ||
+							buttonsDisabled ||
+							executing ||
+							hasPendingChanges,
+						namespace: namespace || "",
+						hasPendingChanges,
+						executing,
+						...commonProps,
+					});
+				case ViewportIconButtonEnum.HistoryMenu:
+					return React.createElement(ButtonComponent, {
+						key: "historyMenu",
+						disabled:
+							!namespace || buttonsDisabled || hasPendingChanges,
+						namespace: namespace || "",
+						visible: iconsVisible,
+						...commonProps,
+					});
+				default:
+					return null;
+			}
+		};
+
+		const sections: React.ReactNode[] = [];
+		if (layout.length === 0) return sections;
+
+		layout.forEach((item, index) => {
+			if (item.type === "button") {
+				const button = renderButtonByKind(
+					item.button.type,
+					componentContext,
+					buttonContext,
+				);
+				if (button) sections.push(button);
+			} else if (item.type === "group") {
+				const groupButtons: React.ReactNode[] = [];
+				item.sections.forEach((section) => {
+					section.forEach((buttonDef) => {
+						const button = renderButtonByKind(
+							buttonDef.type,
+							componentContext,
+							buttonContext,
+						);
+						if (button) groupButtons.push(button);
+					});
+					// Add divider between sections within a group
+					if (
+						groupButtons.length > 0 &&
+						section !== item.sections[item.sections.length - 1]
+					) {
+						groupButtons.push(
+							React.createElement(Divider, {
+								key: `divider-${index}-${section.length}`,
+								...dividerProps,
+							}),
+						);
+					}
+				});
+				sections.push(
+					React.createElement(
+						React.Fragment,
+						{key: `group-${index}`},
+						...groupButtons,
+					),
+				);
+			}
+
+			// Add divider between layout items
+			if (index < layout.length - 1) {
+				sections.push(
+					React.createElement(Divider, {
+						key: `layout-divider-${index}`,
+						...dividerProps,
+					}),
+				);
+			}
+		});
+
+		return sections;
 	}, [layout, componentContext, buttonContext, dividerProps]);
 
 	// Prevent event propagation to avoid triggering viewport interactions
